# Conflicts:
#	TotalCrossVM/builders/build.xml
#	TotalCrossVM/builders/gcc-palm/tcvm/Makefile
#	TotalCrossVM/builders/xcode/tcvm.xcodeproj/project.pbxproj
#	TotalCrossVM/src/event/darwin/event.m
#	TotalCrossVM/src/init/noras_ids/noras.inc
#	TotalCrossVM/src/init/startup.c
#	TotalCrossVM/src/jni/Android.mk

+++ TotalCrossSDK/src/totalcross/net/ssl/SSLCTX4D.java
@@ -51,7 +51,7 @@ public class SSLCTX4D
    /**
     * A reference to the real client/server context. For internal use only.
     */
-   protected int m_ctx;
+   protected long m_ctx;
    
    boolean dontFinalize; //flsobral@tc114_36: finalize support.
